requirements.txt
anyio==3.6.1
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.85.0
h11==0.13.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.0
pycodestyle==2.9.1
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
